Split RS232 up into its own folder with subfiles.
Subfiles include deserializer, serializer, state types,
and ftdi clock functions with type.

The current file organization scheme only allows for
top level files to be in the root of the ./src directory.

The makefile has been updated to support recursively
searching for source files in ./src.

The README has been updated to reflect the fact that
the previous serial to LED example now has its own topfile
distinct from the implementation of its subcomponents. The
name of this topfile affects the invocation of make -
necessitating a update to the instructions in the README.